+++ app/src/main/res/layout/activity_main.xml
@@ -4,4 +4,4 @@
     android:id="@+id/fragment_container_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:name="com.michaeltroger.gruenerpass.view.MainFragment"/>
+    android:name="com.michaeltroger.gruenerpass.MainFragment"/>
